refactor: 섹션 파일 폴더 구조 개선
