+++ src/pages/Home/components/CategorySection/CategorySection.style.ts
