+++ src/pages/Home/components/SelectFriendSection/SelectFriendSection.style.ts
