+++ src/pages/OrderPage/components/GetterInfoSection/GetterInfoSection.style.ts
