+++ src/pages/OrderPage/components/GroupGettersInfoSection/GroupGettersInfoSection.style.ts
