+++ src/pages/OrderPage/components/SenderInfoSection/SenderInfoSection.style.ts
